+++ packages/mobile/src/screens/update-required-screen/index.ts
@@ -1,0 +1,2 @@
+export { RestartRequiredScreen } from './RestartRequiredScreen'
+export { UpdateRequiredScreen } from './UpdateRequiredScreen'
